+++ assemblyadapter-pager2/src/main/java/com/github/panpf/assemblyadapter/pager2/AssemblySingleDataFragmentStateListAdapter.kt
@@ -163,14 +163,14 @@ open class AssemblySingleDataFragmentStateListAdapter<DATA : Any> : FragmentStat
     ) : this(fragment.childFragmentManager, fragment.lifecycle, itemFactory, initData, config)
 
     override fun submitList(list: List<DATA>?) {
-        require(list?.size ?: 0 > 1) {
+        require(list?.size ?: 0 <= 1) {
             "Cannot submit a list with size greater than 1"
         }
         super.submitList(list)
     }
 
     override fun submitList(list: List<DATA>?, commitCallback: Runnable?) {
-        require(list?.size ?: 0 > 1) {
+        require(list?.size ?: 0 <= 1) {
             "Cannot submit a list with size greater than 1"
         }
         super.submitList(list, commitCallback)